354c800c Fixed conflict with new Python wrapping
d10b8807 Fix signed-unsigned comparision warning from dashboard
214c8b1c Remove dead code

+++ ParaViewCore/VTKExtensions/CosmoTools/vtkPGenericIOMultiBlockReader.cxx
@@ -734,7 +734,7 @@ void vtkPGenericIOMultiBlockReader::LoadDataArraysForBlock(vtkUnstructuredGrid *
   block_t& dataBlock = this->MetaData->Blocks[blockId];
 
   assert("pre: # points in dataset different from points in block" &&
-    (grid->GetNumberOfPoints() == dataBlock.NumberOfElements));
+    (static_cast<uint64_t>(grid->GetNumberOfPoints()) == dataBlock.NumberOfElements));
   vtkPointData* PD = grid->GetPointData();
 
   int arrayIdx = 0;
